Added 'open-event-channel' fn to transaction
Promises returned from add and put method of
objectStore sometimes never be called, because
the completion of an transaction should be checked
through events of transaction objects.

open-event-channel returns an channel which
returns an success boxed object if the 'complete'
event is called on transaction object,
and returns an failure boxed object if the 'abort'
or 'error' event occurred.

Users can check the failure box if it is 'abort' or
'error' by checking the :type field of a returned
map data.